(RSP) MIPS offsets are always sign-extended in memory loads/stores.
Part of the purpose of these commits is to get non-portable types (like DWORD) to fixed-size, more readable types that conform to MIPS size types.  While the usage of Windows API types was a part of the problem, it also is not portable to repeatedly say (short)(union.offset) every single time throughout the source, when the MIPS instruction set architecture automatically sign-extends the 16-bit offset value anyway regardless when doing any of these memory loads and stores.  As such, it was counter-intuitive to repeatedly use type conversions to the (short) type, when (short) is not portable either and not necessarily a 16-bit type.  Now that the `offset` member of the RSP opcode union is already re-defined a int16_t (actually a "signed" 16-bit struct bitfield, which is even better), it seems more intuitive to just go with that universally for MIPS offset decoding.  "Find All References" in VS2013 did not detect any instances of referring to `offset` without an explicit type-cast to (short) with only some inconsistent exceptions near the end of the RSP commands debugger that sprintf'd it as a %04X.

I did not make this change outside of the RSP plugin to the main Project64 R4300 CPU source, just the RSP for now.  In a future commit, this change should probably be applied to the main Project64 CPU as well.

+++ Source/RSP/Interpreter Ops.c
@@ -64,7 +64,7 @@ void RSP_Opcode_JAL ( void ) {
 void RSP_Opcode_BEQ ( void ) {
 	RSP_NextInstruction = DELAY_SLOT;
 	if (RSP_GPR[RSPOpC.rs].W == RSP_GPR[RSPOpC.rt].W) {
-		RSP_JumpTo = ( *PrgCount + ((short)RSPOpC.offset << 2) + 4 ) & 0xFFC;
+		RSP_JumpTo = ( *PrgCount + (RSPOpC.offset << 2) + 4 ) & 0xFFC;
 	} else  {
 		RSP_JumpTo = ( *PrgCount + 8 ) & 0xFFC;
 	}
@@ -73,7 +73,7 @@ void RSP_Opcode_BEQ ( void ) {
 void RSP_Opcode_BNE ( void ) {
 	RSP_NextInstruction = DELAY_SLOT;
 	if (RSP_GPR[RSPOpC.rs].W != RSP_GPR[RSPOpC.rt].W) {
-		RSP_JumpTo = ( *PrgCount + ((short)RSPOpC.offset << 2) + 4 ) & 0xFFC;
+		RSP_JumpTo = ( *PrgCount + (RSPOpC.offset << 2) + 4 ) & 0xFFC;
 	} else  {
 		RSP_JumpTo = ( *PrgCount + 8 ) & 0xFFC;
 	}
@@ -82,7 +82,7 @@ void RSP_Opcode_BNE ( void ) {
 void RSP_Opcode_BLEZ ( void ) {
 	RSP_NextInstruction = DELAY_SLOT;
 	if (RSP_GPR[RSPOpC.rs].W <= 0) {
-		RSP_JumpTo = ( *PrgCount + ((short)RSPOpC.offset << 2) + 4 ) & 0xFFC;
+		RSP_JumpTo = ( *PrgCount + (RSPOpC.offset << 2) + 4 ) & 0xFFC;
 	} else  {
 		RSP_JumpTo = ( *PrgCount + 8 ) & 0xFFC;
 	}
@@ -91,7 +91,7 @@ void RSP_Opcode_BLEZ ( void ) {
 void RSP_Opcode_BGTZ ( void ) {
 	RSP_NextInstruction = DELAY_SLOT;
 	if (RSP_GPR[RSPOpC.rs].W > 0) {
-		RSP_JumpTo = ( *PrgCount + ((short)RSPOpC.offset << 2) + 4 ) & 0xFFC;
+		RSP_JumpTo = ( *PrgCount + (RSPOpC.offset << 2) + 4 ) & 0xFFC;
 	} else  {
 		RSP_JumpTo = ( *PrgCount + 8 ) & 0xFFC;
 	}
@@ -147,7 +147,7 @@ void RSP_Opcode_XORI ( void ) {
 
 void RSP_Opcode_LUI (void) {
 	if (RSPOpC.rt != 0) {
-		RSP_GPR[RSPOpC.rt].W = (short)RSPOpC.offset << 16;
+		RSP_GPR[RSPOpC.rt].W = RSPOpC.offset << 16;
 	}
 }
 
@@ -160,46 +160,46 @@ void RSP_Opcode_COP2 (void) {
 }
 
 void RSP_Opcode_LB ( void ) {
-	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + (short)RSPOpC.offset) & 0xFFF);
+	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + RSPOpC.offset) & 0xFFF);
 	RSP_LB_DMEM( Address, &RSP_GPR[RSPOpC.rt].UB[0] );
 	RSP_GPR[RSPOpC.rt].W = RSP_GPR[RSPOpC.rt].B[0];
 }
 
 void RSP_Opcode_LH ( void ) {
-	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + (short)RSPOpC.offset) & 0xFFF);
+	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + RSPOpC.offset) & 0xFFF);
 	RSP_LH_DMEM( Address, &RSP_GPR[RSPOpC.rt].UHW[0] );
 	RSP_GPR[RSPOpC.rt].W = RSP_GPR[RSPOpC.rt].HW[0];
 }
 
 void RSP_Opcode_LW ( void ) {
-	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + (short)RSPOpC.offset) & 0xFFF);
+	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + RSPOpC.offset) & 0xFFF);
 	RSP_LW_DMEM( Address, &RSP_GPR[RSPOpC.rt].UW );
 }
 
 void RSP_Opcode_LBU ( void ) {
-	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + (short)RSPOpC.offset) & 0xFFF);
+	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + RSPOpC.offset) & 0xFFF);
 	RSP_LB_DMEM( Address, &RSP_GPR[RSPOpC.rt].UB[0] );
 	RSP_GPR[RSPOpC.rt].UW = RSP_GPR[RSPOpC.rt].UB[0];
 }
 
 void RSP_Opcode_LHU ( void ) {
-	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + (short)RSPOpC.offset) & 0xFFF);
+	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + RSPOpC.offset) & 0xFFF);
 	RSP_LH_DMEM( Address, &RSP_GPR[RSPOpC.rt].UHW[0] );
 	RSP_GPR[RSPOpC.rt].UW = RSP_GPR[RSPOpC.rt].UHW[0];
 }
 
 void RSP_Opcode_SB ( void ) {
-	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + (short)RSPOpC.offset) & 0xFFF);
+	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + RSPOpC.offset) & 0xFFF);
 	RSP_SB_DMEM( Address, RSP_GPR[RSPOpC.rt].UB[0] );
 }
 
 void RSP_Opcode_SH ( void ) {
-	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + (short)RSPOpC.offset) & 0xFFF);
+	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + RSPOpC.offset) & 0xFFF);
 	RSP_SH_DMEM( Address, RSP_GPR[RSPOpC.rt].UHW[0] );
 }
 
 void RSP_Opcode_SW ( void ) {
-	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + (short)RSPOpC.offset) & 0xFFF);
+	DWORD Address = ((RSP_GPR[RSPOpC.base].UW + RSPOpC.offset) & 0xFFF);
 	RSP_SW_DMEM( Address, RSP_GPR[RSPOpC.rt].UW );
 }
 
@@ -336,7 +336,7 @@ void RSP_Special_SLTU (void) {
 void RSP_Opcode_BLTZ ( void ) {
 	RSP_NextInstruction = DELAY_SLOT;
 	if (RSP_GPR[RSPOpC.rs].W < 0) {
-		RSP_JumpTo = ( *PrgCount + ((short)RSPOpC.offset << 2) + 4 ) & 0xFFC;
+		RSP_JumpTo = ( *PrgCount + (RSPOpC.offset << 2) + 4 ) & 0xFFC;
 	} else  {
 		RSP_JumpTo = ( *PrgCount + 8 ) & 0xFFC;
 	}
@@ -345,7 +345,7 @@ void RSP_Opcode_BLTZ ( void ) {
 void RSP_Opcode_BGEZ ( void ) {
 	RSP_NextInstruction = DELAY_SLOT;
 	if (RSP_GPR[RSPOpC.rs].W >= 0) {
-		RSP_JumpTo = ( *PrgCount + ((short)RSPOpC.offset << 2) + 4 ) & 0xFFC;
+		RSP_JumpTo = ( *PrgCount + (RSPOpC.offset << 2) + 4 ) & 0xFFC;
 	} else  {
 		RSP_JumpTo = ( *PrgCount + 8 ) & 0xFFC;
 	}
@@ -355,7 +355,7 @@ void RSP_Opcode_BLTZAL ( void ) {
 	RSP_NextInstruction = DELAY_SLOT;
 	RSP_GPR[31].UW = ( *PrgCount + 8 ) & 0xFFC;
 	if (RSP_GPR[RSPOpC.rs].W < 0) {
-		RSP_JumpTo = ( *PrgCount + ((short)RSPOpC.offset << 2) + 4 ) & 0xFFC;
+		RSP_JumpTo = ( *PrgCount + (RSPOpC.offset << 2) + 4 ) & 0xFFC;
 	} else  {
 		RSP_JumpTo = ( *PrgCount + 8 ) & 0xFFC;
 	}
@@ -365,7 +365,7 @@ void RSP_Opcode_BGEZAL ( void ) {
 	RSP_NextInstruction = DELAY_SLOT;
 	RSP_GPR[31].UW = ( *PrgCount + 8 ) & 0xFFC;
 	if (RSP_GPR[RSPOpC.rs].W >= 0) {
-		RSP_JumpTo = ( *PrgCount + ((short)RSPOpC.offset << 2) + 4 ) & 0xFFC;
+		RSP_JumpTo = ( *PrgCount + (RSPOpC.offset << 2) + 4 ) & 0xFFC;
 	} else  {
 		RSP_JumpTo = ( *PrgCount + 8 ) & 0xFFC;
 	}
